+++ lib/moderation.rb
@@ -29,11 +29,11 @@ module Moderation
       @constructor    = options.fetch(:constructor, :no_constructor)
       @construct_with = options.fetch(:construct_with, :no_construct_with)
       @limit          = options.fetch(:limit, Store.configuration.limit)
-      @storage        = options.fetch(:storage) { Adapters::MemoryAdapter.new }
       @coercer        = options.fetch(:coercer) do
         require_relative 'moderation/coercer/multi_json_coercer'
         Coercer::MultiJsonCoercer.new
       end
+      @storage        = options.fetch(:storage) { Adapters::MemoryAdapter.new(coercer: @coercer) }
       @storage.limit  = @limit
     end